+++ cli/cli/commands/service/logs/logs.go
@@ -14,6 +14,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/api/golang/engine/lib/kurtosis_context"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/highlevel/enclave_id_arg"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/highlevel/engine_consuming_kurtosis_command"
+	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/highlevel/service_guid_arg"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/lowlevel/args"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/lowlevel/flags"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_str_consts"
@@ -35,7 +36,9 @@ const (
 	isEnclaveIdArgOptional = false
 	isEnclaveIdArgGreedy   = false
 
-	serviceGuidArgKey = "service-guid"
+	serviceGuidArgKey        = "service-guid"
+	isServiceGuidArgOptional = false
+	isServiceGuidArgGreedy   = false
 
 	shouldFollowLogsFlagKey = "follow"
 
@@ -69,10 +72,11 @@ var ServiceLogsCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosisC
 			isEnclaveIdArgOptional,
 			isEnclaveIdArgGreedy,
 		),
-		// TODO Create a NewServiceIDArg that adds autocomplete
-		{
-			Key: serviceGuidArgKey,
-		},
+		service_guid_arg.NewServiceGUIDArg(
+			serviceGuidArgKey,
+			isServiceGuidArgOptional,
+			isServiceGuidArgGreedy,
+		),
 	},
 	RunFunc: run,
 }